Legal_Agent/chatbot/requirements.txt
fastapi<=0.95.2
uvicorn<=0.22.0
openai<=1.1.1
pinecone-client==2.2.1
python-dotenv==0.21.1
pydantic<=1.10.13
pymysql<=1.1.0
python-jose<=3.3.0
python-multipart<=0.0.6
anthropic<=0.5.0